neo-core2/src/crypto/keys/wif_test.rs
use hex;
use base58;
use std::error::Error;
use std::fmt;
use std::str;
use std::string::ToString;

#[derive(Debug)]
struct WIFError;

impl fmt::Display for WIFError {
    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
        write!(f, "WIF Error")
    }
}

impl Error for WIFError {}

struct WIF {
    private_key: Vec<u8>,
    compressed: bool,
    version: u8,
}

impl WIF {
    fn to_string(&self) -> String {
        hex::encode(&self.private_key)
    }
}

fn wif_encode(private_key: &[u8], version: u8, compressed: bool) -> Result<String, Box<dyn Error>> {
    // Implement WIF encoding logic here
    Ok(String::new())
}

fn wif_decode(wif: &str, version: u8) -> Result<WIF, Box<dyn Error>> {
    // Implement WIF decoding logic here
    Ok(WIF {
        private_key: vec![],
        compressed: false,
        version,
    })
}

struct WifTestCase {
    wif: String,
    compressed: bool,
    private_key: String,
    version: u8,
}

fn get_wif_test_cases() -> Vec<WifTestCase> {
    vec![
        WifTestCase {
            wif: "KwDiBf89QgGbjEhKnhXJuH7LrciVrZi3qYjgd9M7rFU73sVHnoWn".to_string(),
            compressed: true,
            private_key: "0000000000000000000000000000000000000000000000000000000000000001".to_string(),
            version: 0x80,
        },
        WifTestCase {
            wif: "5HpHagT65TZzG1PH3CSu63k8DbpvD8s5ip4nEB3kEsreAnchuDf".to_string(),
            compressed: false,
            private_key: "0000000000000000000000000000000000000000000000000000000000000001".to_string(),
            version: 0x80,
        },
        WifTestCase {
            wif: "KxhEDBQyyEFymvfJD96q8stMbJMbZUb6D1PmXqBWZDU2WvbvVs9o".to_string(),
            compressed: true,
            private_key: "2bfe58ab6d9fd575bdc3a624e4825dd2b375d64ac033fbc46ea79dbab4f69a3e".to_string(),
            version: 0x80,
        },
        WifTestCase {
            wif: "KxhEDBQyyEFymvfJD96q8stMbJMbZUb6D1PmXqBWZDU2WvbvVs9o".to_string(),
            compressed: true,
            private_key: "2bfe58ab6d9fd575bdc3a624e4825dd2b375d64ac033fbc46ea79dbab4f69a3e".to_string(),
            version: 0x00,
        },
    ]
}

#[test]
fn test_wif_encode_decode() {
    let wif_test_cases = get_wif_test_cases();
    for test_case in wif_test_cases {
        let b = hex::decode(&test_case.private_key).unwrap();
        let wif = wif_encode(&b, test_case.version, test_case.compressed).unwrap();
        assert_eq!(test_case.wif, wif);

        let wif_decoded = wif_decode(&wif, test_case.version).unwrap();
        assert_eq!(test_case.private_key, wif_decoded.to_string());
        assert_eq!(test_case.compressed, wif_decoded.compressed);
        if test_case.version != 0 {
            assert_eq!(test_case.version, wif_decoded.version);
        } else {
            assert_eq!(0x80, wif_decoded.version); // Assuming WIFVersion is 0x80
        }
    }

    let wif_inv = vec![0, 1, 2];
    let result = wif_encode(&wif_inv, 0, true);
    assert!(result.is_err());
}

#[test]
fn test_bad_wif_decode() {
    let result = wif_decode("garbage", 0);
    assert!(result.is_err());

    let s = base58::check_encode(&[]);
    let result = wif_decode(&s, 0);
    assert!(result.is_err());

    let mut uncompr = vec![0; 33];
    let mut compr = vec![0; 34];

    let s = base58::check_encode(&compr);
    let result = wif_decode(&s, 0);
    assert!(result.is_err());

    let s = base58::check_encode(&uncompr);
    let result = wif_decode(&s, 0);
    assert!(result.is_err());

    compr[33] = 1;
    compr[0] = 0x80; // Assuming WIFVersion is 0x80
    uncompr[0] = 0x80; // Assuming WIFVersion is 0x80

    let s = base58::check_encode(&compr);
    let result = wif_decode(&s, 0);
    assert!(result.is_ok());

    let s = base58::check_encode(&uncompr);
    let result = wif_decode(&s, 0);
    assert!(result.is_ok());
}
